u32's make clang happy

+++ yave/DescriptorSetLayoutPool.cpp
@@ -33,7 +33,7 @@ vk::DescriptorSetLayout DescriptorSetLayoutPool::create_descriptor_set_layout(co
 	auto& layout = _layouts[bindings];
 	if(!layout) {
 		layout = get_device()->get_vk_device().createDescriptorSetLayout(vk::DescriptorSetLayoutCreateInfo()
-				.setBindingCount(bindings.size())
+				.setBindingCount(u32(bindings.size()))
 				.setPBindings(bindings.begin())
 			);
 	}
